Add `text-dark-mode` to `select` and `input` tags
When this change is applied, currently existing `select` tags will be
styled using `text-dark-mode` to match the rest of the text in the
application. This is significantly more readable in the default theme.

The same style is also applied to text inputs in the config panel. i.e.
Chat Notifier config.